backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
supabase
requests
selenium
undetected-chromedriver
beautifulsoup4
pandas
